+++ libfabric.vcxproj
@@ -571,7 +571,7 @@
     <ClInclude Include="include\ofi_mr.h" />
     <ClInclude Include="include\ofi_net.h" />
     <ClInclude Include="include\fi_enosys.h" />
-    <ClInclude Include="include\fi_file.h" />
+    <ClInclude Include="include\ofi_file.h" />
     <ClInclude Include="include\fi_indexer.h" />
     <ClInclude Include="include\fi_list.h" />
     <ClInclude Include="include\fi_lock.h" />
