+++ libfabric.vcxproj.filters
@@ -476,7 +476,7 @@
     <ClInclude Include="include\fi_atom.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="include\fi_file.h">
+    <ClInclude Include="include\ofi_file.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="include\fi_lock.h">
