+++ prov/gni/include/gnix.h
@@ -61,7 +61,7 @@
 #include <fi_enosys.h>
 #include <fi_rbuf.h>
 #include <fi_list.h>
-#include <fi_file.h>
+#include <ofi_file.h>
 
 #ifdef HAVE_UDREG
 #include <udreg_pub.h>
