+++ prov/sockets/src/sock_conn.c
@@ -53,7 +53,7 @@
 
 #include "sock.h"
 #include "sock_util.h"
-#include "fi_file.h"
+#include "ofi_file.h"
 
 #define SOCK_LOG_DBG(...) _SOCK_LOG_DBG(FI_LOG_EP_CTRL, __VA_ARGS__)
 #define SOCK_LOG_ERROR(...) _SOCK_LOG_ERROR(FI_LOG_EP_CTRL, __VA_ARGS__)
