+++ prov/usnic/src/usdf_cm.c
@@ -54,7 +54,7 @@
 #include <rdma/fi_rma.h>
 #include <rdma/fi_errno.h>
 #include "fi.h"
-#include "fi_file.h"
+#include "ofi_file.h"
 
 #include "usnic_direct.h"
 #include "usdf.h"
