+++ prov/verbs/src/ep_rdm/verbs_ep_rdm.c
@@ -36,7 +36,7 @@
 #include <rdma/rdma_cma.h>
 
 #include <fi_list.h>
-#include <fi_file.h>
+#include <ofi_file.h>
 #include <fi_enosys.h>
 #include "../fi_verbs.h"
 #include "verbs_queuing.h"
