+++ src/unix/osd.c
@@ -50,7 +50,7 @@
 
 #include "fi.h"
 #include "fi_osd.h"
-#include "fi_file.h"
+#include "ofi_file.h"
 #include "fi_util.h"
 
 #include "rdma/fi_errno.h"
